backend/requirements.txt
fastapi
uvicorn
ib_insync
openai
psycopg2-binary
python-dotenv
sqlalchemy
alembic
nest_asyncio
yfinance 